+++ src/Models/delivery/Transformers/ShippingAddonTransformer.ts
@@ -2,7 +2,7 @@ import ShippingAddon from '../ShippingAddon';
 import ShippingAddonInterface from '../../../Interfaces/delivery/ShippingAddon';
 import BaseTransformer from '../../../BaseTransformer';
 
-export default class GiftTransformer extends BaseTransformer {
+export default class ShippingAddonTransformer extends BaseTransformer {
     /**
      * Map an shipping addon response
      *